Renaming of multi generator file
