+++ lib/multi-database-9000.rb
@@ -1,6 +1,6 @@
 require "multi-database-9000/version"
 require "multi-database-9000/multi-database-9000"
-require "multi-database-9000/generators/migration_generator"
+require "multi-database-9000/generators/multi_migration_generator"
 
 module MultiDatabase9000
   class Railtie < Rails::Railtie
