+++ lib/multi-database-9000/generators/multi_migration_generator.rb
